Add files via upload
New commands added:
!shop and !buy
Also, when using these commands with different conditions, errors are displayed in the console. Also, for each error, the time at which it occurred is shown. Over time, with each command in the code the error and its time will be shown

UPD: To create a new database. Just delete mine. Launching the bot via "start.bat"